BugFix And New Levels
Bug fix is that before, when the player was out of lives and died and pressed restart, it would show the ad and then once skiped would restart the game from level 1 but it would be forzen. That was fixed but then I saw that the level counter at the top of the screen would not go back to level 1 and at the time I did not know how to access static veriables using other scripts, (which was vital in this case because the level number was a public static int in the next level ai script and had to be accesed in the ad manager script). Once I saw a youtube tutorial, everthing was cleared up by just using NextLevelAI.LevelNumber = 0